requirements.txt
autopep8==1.5.4
click==7.1.2
easydict==1.9
fastapi==0.63.0
h11==0.12.0
mongoengine==0.22.1
motor==1.2.1
motorengine==1.0.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pycodestyle==2.6.0
pydantic==1.7.3
pymongo==3.6.0
six==1.15.0
starlette==0.13.6
toml==0.10.2
tornado==6.1
uvicorn==0.13.3
